+++ plugins/source.go
@@ -46,6 +46,9 @@ const (
 // Add internal columns
 func addInternalColumns(tables []*schema.Table) {
 	for _, table := range tables {
+		if c := table.Column("_cq_id"); c != nil {
+			continue
+		}
 		cqID := schema.CqIDColumn
 		if len(table.PrimaryKeys()) == 0 {
 			cqID.CreationOptions.PrimaryKey = true