+++ plugins/source_validate_test.go
@@ -32,6 +32,21 @@ func TestSourcePlugin_listAndValidateAllResources(t *testing.T) {
 			wantErr:                 false,
 		},
 		{
+			name:                    "should return error when invalid table is skipped",
+			plugin:                  SourcePlugin{tables: []*schema.Table{{Name: "table 1"}, {Name: "table 2"}, {Name: "table 3"}}},
+			configurationTables:     []string{"*"},
+			configurationSkipTables: []string{"table 4"},
+			wantErr:                 true,
+		},
+		{
+			name:                    "should return an error if child table is skipped",
+			plugin:                  SourcePlugin{tables: []*schema.Table{{Name: "main_table", Relations: []*schema.Table{{Name: "sub_table", Parent: &schema.Table{Name: "main_table"}}}}}},
+			configurationTables:     []string{"*"},
+			configurationSkipTables: []string{"sub_table"},
+			wantErr:                 true,
+		},
+
+		{
 			name:                "should return specific tables when they are provided",
 			plugin:              SourcePlugin{tables: []*schema.Table{{Name: "table 1"}, {Name: "table 2"}, {Name: "table 3"}}},
 			configurationTables: []string{"table 1"},